requirements.txt
requests
python-dotenv
pandas
seaborn
matplotlib
scikit-learn
numpy
scipy